requirements.txt
Flask
Session
werkzeug.security
